Mint ERC721 via smartcontract
Update to previous prefab. renaming for clarity.
