+++ Web3Unity/Prefabs/Wallet/Transfer721MintContract.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c2827d5b8bd994fa58030f130d48236f
+guid: 102be22631d01934dacb5c89ebf02394
 PrefabImporter:
   externalObjects: {}
   userData:
